pyproject.toml
[tool.poetry]
name = "Quiz Engine"
version = "0.1.0"
description = ""
authors = ["Kabeer Vohra <kabeer@vohra.eu>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.63.0"
SQLAlchemy = "1.3.19"
uvicorn = "0.13.3"
starlette = "0.13.6"
psycopg2-binary = "^2.9.3"
alembic = "^1.8.1"
pydantic = "^1.9.1"
passlib = "^1.7.4"
python-jose = "^3.3.0"
python-multipart = "^0.0.5"
pre-commit = "^2.20.0"
pytest = "^7.1.2"
"testing.postgresql" = "^1.3.0"
mock = "^4.0.3"

[tool.poetry.dev-dependencies]
pudb = "^2022.1.2"
jedi = "^0.18.1"
pynvim = "^0.4.3"
